requirements.txt
# Core runtime dependencies
Flask>=3.0.0
requests>=2.31.0
delta-rest-client>=1.0.13
pandas>=2.2.0
websockets>=13.0.0